feat(dprint): overhaul dprint integration
