+++ .github/actions/setup/action.yml
@@ -6,6 +6,10 @@ inputs:
     description: Where the package files are located
     default: ""
     required: false
+  setup-dprint:
+    description: Whether to set up dprint as well
+    default: "false"
+    required: false
 
 runs:
   using: composite
@@ -16,13 +20,30 @@ runs:
 
     - uses: actions/setup-node@v4
       with:
-        node-version: "23.0.0-nightly"
+        node-version: "23.11.0"
         cache: "pnpm"
         cache-dependency-path: ${{ inputs.root }}pnpm-lock.yaml
 
     - name: install
       working-directory: ${{ inputs.root }}
+      shell: bash
       run: pnpm install --frozen-lockfile
+
+    - id: dprint_config
+      name: read dprint config
+      if: ${{ inputs.install-dprint }}
+      working-directory: ${{ inputs.root }}
+      run: |
+        content="cat ./dprint.json"
+        echo "::set-output name=config::$content"
+      shell: bash
+
+    - name: install dprint
+      if: ${{ inputs.install-dprint }}
+      working-directory: ${{ inputs.root }}
+      run: |
+        curl -fsSL https://dprint.dev/install.sh | sh -s ${{ fromJson(steps.dprint_config.outputs.config).version }} > /dev/null 2>&1
+        echo "/home/runner/.dprint/bin" >> $GITHUB_PATH
       shell: bash
 
     - run: |
