+++ docs/README.md
@@ -12,8 +12,4 @@
 
 ### For developers
 
-- ### [Local building](./BUILDING.md)
-
-  - Learn how to build RevengePlugins locally and serve it to your phone
-
 - ### Themes+ — [Making your custom icons themable](./ICONS_DEV_HOW2.md)
